vertebrae/core: Moved most internal functions outside of script
...and created vertebrae/core.py
